requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
jinja2>=3.1.0
python-multipart>=0.0.6
pandas>=2.0.0
#weasyprint>=64.0
psycopg2-binary>=2.9.0
alembic
python-dotenv>=1.0.0
